requirements.txt
kbcstorage>=0.9.2
qdrant-client>=1.7.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
msgpack>=1.0.5
